add logging and rename duckdb services
